+++ Editor/Profiler/ObjectData.cs
@@ -54,7 +54,7 @@ namespace Unity.MemoryProfiler.Editor
         {
             get
             {
-                return CachedSnapshot.NativeObjectEntriesCache.InstanceID_None;
+                return CachedSnapshot.NativeObjectEntriesCache.k_InstanceIDNone;
             }
         }
         private ObjectDataType m_dataType;
@@ -472,7 +472,7 @@ namespace Unity.MemoryProfiler.Editor
             {
                 return snapshot.nativeObjects.instanceId[nativeIndex];
             }
-            return CachedSnapshot.NativeObjectEntriesCache.InstanceID_None;
+            return CachedSnapshot.NativeObjectEntriesCache.k_InstanceIDNone;
         }
 
         public ObjectData GetBase(CachedSnapshot snapshot)
@@ -507,13 +507,11 @@ namespace Unity.MemoryProfiler.Editor
                 case ObjectDataType.Object:
                 case ObjectDataType.BoxedValue:
                 {
-                    ManagedObjectInfo moi;
-                    if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(m_data.managed.objectPtr, out moi))
+                    int idx;
+                    if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(m_data.managed.objectPtr, out idx))
                     {
-                        return snapshot.ManagedObjectIndexToUnifiedObjectIndex(moi.ManagedObjectIndex);
+                        return snapshot.ManagedObjectIndexToUnifiedObjectIndex(idx);
                     }
-
-
                     break;
                 }
                 case ObjectDataType.NativeObject:
@@ -531,23 +529,23 @@ namespace Unity.MemoryProfiler.Editor
                 case ObjectDataType.Object:
                 case ObjectDataType.BoxedValue:
                     {
-                        ManagedObjectInfo moi;
-                        if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(m_data.managed.objectPtr, out moi))
+                        int idx;
+                        if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(m_data.managed.objectPtr, out idx))
                         {
-                            return moi;
+                            return snapshot.CrawledData.ManagedObjects[idx];
                         }
                         throw new Exception("Invalid object pointer used to query object list.");
                     }
                 case ObjectDataType.ReferenceObject:
                 case ObjectDataType.ReferenceArray:
                     {
-                        ManagedObjectInfo moi;
+                        int idx;
                         ulong refPtr = GetReferencePointer();
                         if (refPtr == 0)
                             return default(ManagedObjectInfo);
-                        if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(GetReferencePointer(), out moi))
+                        if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(GetReferencePointer(), out idx))
                         {
-                            return moi;
+                            return snapshot.CrawledData.ManagedObjects[idx];
                         }
                         throw new Exception("Invalid object reference pointer used to query object list.");
                     }
@@ -564,10 +562,10 @@ namespace Unity.MemoryProfiler.Editor
                 case ObjectDataType.Object:
                 case ObjectDataType.BoxedValue:
                 {
-                    ManagedObjectInfo moi;
-                    if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(m_data.managed.objectPtr, out moi))
+                    int idx;
+                    if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(m_data.managed.objectPtr, out idx))
                     {
-                        return moi.ManagedObjectIndex;
+                        return idx;
                     }
 
 
@@ -667,6 +665,11 @@ namespace Unity.MemoryProfiler.Editor
 
             if (index < snapshot.gcHandles.Count)
             {
+                //When snapshotting we might end up getting some handle targets as they are about to be collected
+                //we do restart the world temporarily this can cause us to end up with targets that are not present in the dumped heaps
+                if (moi.PtrObject == 0)
+                    return ObjectData.invalid;
+
                 if (moi.PtrObject != snapshot.gcHandles.target[index])
                 {
                     throw new Exception("bad object");
@@ -679,10 +682,10 @@ namespace Unity.MemoryProfiler.Editor
         public static ObjectData FromManagedPointer(CachedSnapshot snapshot, ulong ptr, int asTypeIndex = -1)
         {
             if (ptr == 0) return ObjectData.invalid;
-            ManagedObjectInfo moi;
-            if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(ptr, out moi))
+            int idx;
+            if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(ptr, out idx))
             {
-                return FromManagedObjectInfo(snapshot, moi);
+                return FromManagedObjectInfo(snapshot, snapshot.CrawledData.ManagedObjects[idx]);
             }
             else
             {
@@ -774,10 +777,10 @@ namespace Unity.MemoryProfiler.Editor
                 case ObjectDataType.BoxedValue:
                 case ObjectDataType.Object:
                 {
-                    ManagedObjectInfo moi;
-                    if (snapshot.CrawledData.ManagedObjectByAddress.TryGetValue(obj.hostManagedObjectPtr, out moi))
+                    int idx;
+                    if (snapshot.CrawledData.MangedObjectIndexByAddress.TryGetValue(obj.hostManagedObjectPtr, out idx))
                     {
-                        objIndex = snapshot.ManagedObjectIndexToUnifiedObjectIndex(moi.ManagedObjectIndex);
+                        objIndex = snapshot.ManagedObjectIndexToUnifiedObjectIndex(idx);
 
                         //add crawled connections
                         for (int i = 0; i != snapshot.CrawledData.Connections.Count; ++i)
@@ -786,13 +789,13 @@ namespace Unity.MemoryProfiler.Editor
                             switch (c.connectionType)
                             {
                                 case ManagedConnection.ConnectionType.Global_To_ManagedObject:
-                                    if (c.toManagedObjectIndex == moi.ManagedObjectIndex)
+                                    if (c.toManagedObjectIndex == idx)
                                     {
                                         o.Add(ObjectData.global);
                                     }
                                     break;
                                 case ManagedConnection.ConnectionType.ManagedObject_To_ManagedObject:
-                                    if (c.toManagedObjectIndex == moi.ManagedObjectIndex)
+                                    if (c.toManagedObjectIndex == idx)
                                     {
                                         var objParent = ObjectData.FromManagedObjectIndex(snapshot, c.fromManagedObjectIndex);
                                         if (c.fieldFrom >= 0)
@@ -810,7 +813,7 @@ namespace Unity.MemoryProfiler.Editor
                                     }
                                     break;
                                 case ManagedConnection.ConnectionType.ManagedType_To_ManagedObject:
-                                    if (c.toManagedObjectIndex == moi.ManagedObjectIndex)
+                                    if (c.toManagedObjectIndex == idx)
                                     {
                                         var objType = ObjectData.FromManagedType(snapshot, c.fromManagedType);
                                         if (c.fieldFrom >= 0)
@@ -828,7 +831,7 @@ namespace Unity.MemoryProfiler.Editor
                                     }
                                     break;
                                 case ManagedConnection.ConnectionType.UnityEngineObject:
-                                    if (c.UnityEngineManagedObjectIndex == moi.ManagedObjectIndex)
+                                    if (c.UnityEngineManagedObjectIndex == idx)
                                     {
                                         o.Add(ObjectData.FromNativeObjectIndex(snapshot, c.UnityEngineNativeObjectIndex));
                                     }